+++ internal/app/storage/models/user_balance.go
@@ -3,5 +3,5 @@ package models
 type UserBalance struct {
 	ID        int     `json:"id"`
 	Balance   float32 `json:"balance"`
-	Withdrawn int     `json:"withdrawn"`
+	Withdrawn float32 `json:"withdrawn"`
 }
